refactor: ArticleContent와 동일하게 Comment관련 된 컴포넌트를 묶은 하나의 컴포넌트로 분리 및 커스텀 훅 적용
